fix all func in persegi, kubus

+++ src/benda_tiga_dimensi/Kubus.java
@@ -2,10 +2,10 @@ package benda_tiga_dimensi;
 import benda_dua_dimensi.Persegi;
 
 public class Kubus extends Persegi{
-    public float volumeKubus;
-    public float luasPermukaanKubus;
+    public double volumeKubus;
+    public double luasPermukaanKubus;
     
-    public Kubus(float sisi){
+    public Kubus(double sisi){
         super(sisi);
         volumeKubus();
         luasPermukaanKubus();
